refactor(refactored blockstyle code ): refactored block code to reuse the styles
